requirements.txt
sotabencheval
torch
torchvision==0.4.0
tqdm
numpy
scipy
pillow==6.2.0
efficientnet_pytorch
efficientnet
keras==2.2.4
tensorflow-gpu==1.9.0
keras_applications>=1.0.7,<=1.0.8
scikit-image

